move asset_transformer.dart to tools directory
